+++ cmd/sparkwing/main.go
@@ -494,6 +494,8 @@ func runCluster(args []string) error {
 		return runUsers(args[1:])
 	case "tokens":
 		return runTokens(args[1:])
+	case "credits":
+		return runCredits(args[1:])
 	case "image":
 		return runImage(args[1:])
 	case "webhooks":